Moved IsManagedObjectNotFound to vclib package
