+++ pkg/cloudprovider/providers/vsphere/vsphere_util.go
@@ -35,9 +35,6 @@ import (
 
 	"k8s.io/kubernetes/pkg/cloudprovider/providers/vsphere/vclib"
 	"k8s.io/kubernetes/pkg/cloudprovider/providers/vsphere/vclib/diskmanagers"
-
-	"github.com/vmware/govmomi/vim25/soap"
-	"github.com/vmware/govmomi/vim25/types"
 )
 
 const (
@@ -316,12 +313,4 @@ func (vs *VSphere) cleanUpDummyVMs(dummyVMPrefix string) {
 			glog.V(4).Infof("Unable to clean up dummy VM's in the kubernetes cluster: %q. err: %+v", vs.cfg.Global.WorkingDir, err)
 		}
 	}
-}
-
-func IsManagedObjectNotFoundError(err error) bool {
-	isManagedObjectNotFoundError := false
-	if soap.IsSoapFault(err) {
-		_, isManagedObjectNotFoundError = soap.ToSoapFault(err).VimFault().(types.ManagedObjectNotFound)
-	}
-	return isManagedObjectNotFoundError
 }
